requirements.txt
pyserial
matplotlib
numpy
pytest
pytest-mock